requirements.txt
pandas>=2.2
openpyxl>=3.1
pyyaml>=6.0
pytest>=8.0
